+++ R/emmip.R
@@ -56,6 +56,14 @@ emmip = function(object, formula, ...) {
 #'   parameters to \code{\link{predict.emmGrid}} (e.g., \code{sigma}) may be passed via
 #'   \code{...}. For Bayesian models, PIs require \code{frequentist = TRUE} and 
 #'   a value for \code{sigma}.
+#' @param style Optional character value. If \code{"factor"}, the levels of the
+#'   horizontal variable are converted to a factor, resulting in equally-spaced
+#'   plotting positions. If \code{"numeric"}, the horizontal variable is taken
+#'   as numeric. This makes sense only when it is in fact a single numeric variable;
+#'   otherwise, the values will likely be confusing (e.g., it is a factor or a
+#'   combination of variables). If not specified, it is auto-detected.
+#'   \code{style} also affects the appearance of each curve; if \code{"factor"},
+#'   points are added to lines, if \code{"numeric"}, we get lines but no points.
 #' @param engine Character value matching \code{"ggplot"} (default) or 
 #'   \code{"lattice"}. The graphics engine to be used to produce the plot.
 #'   These require, respectively, the \pkg{ggplot2} or \pkg{lattice} package to
@@ -124,9 +132,16 @@ emmip = function(object, formula, ...) {
 #' # Individual traces in panels
 #' emmip(noise.lm, ~ size | type * side)
 #' 
+#' # Example for the 'style' argument
+#'  fib.lm = lm(strength ~ machine * sqrt(diameter), data = fiber)
+#'  fib.rg = ref_grid(fib.lm, at = list(diameter = c(3.5, 4, 4.5, 5, 5.5, 6)^2))
+#'  emmip(fib.rg, machine ~ diameter)   # curves (because diameter is numeric)
+#'  emmip(fib.rg, machine ~ diameter, style = "factor")  # lines
+#' 
 #' # For an example using extra ggplot2 code, see 'vignette("messy-data")',
-#' in the section on nested models.
+#' # in the section on nested models.
 emmip.default = function(object, formula, type, CIs = FALSE, PIs = FALSE,
+                         style,
                          engine = get_emm_option("graphics.engine"),
                          pch = c(1,2,6,7,9,10,15:20), 
                          lty = 1, col = NULL, plotit = TRUE, 
@@ -207,7 +222,25 @@ emmip.default = function(object, formula, type, CIs = FALSE, PIs = FALSE,
     
     xvars = specs$rhs
     xv = do.call(paste, unname(emms[xvars]))
-    emms$xvar = factor(xv, levels = unique(xv))
+    ltest = max(apply(table(xv,tv), 2, function(x) sum(x > 0))) # length of longest trace
+    if (missing(style))
+        style = ifelse(length(xvars) == 1 && 
+                           is.numeric(emms[[xvars]]) &&
+                           ltest > 1,
+                   "numeric", "factor")
+    else
+        style = match.arg(style, c("factor", "numeric"))
+    if (style == "factor") {
+        emms$xvar = factor(xv, levels = unique(xv))
+        predicate = "Levels of "
+        if (ltest <= 1)
+            message("Suggestion: Add 'at = list(", xvars, " = ...)' ",
+                    "to call to see > 1 value per group.")
+    }
+    else {
+        emms$xvar = as.numeric(xv)
+        predicate = ""
+    }
     emms = emms[order(emms$xvar), ]
     plotform = yvar ~ xvar
     
@@ -216,7 +249,7 @@ emmip.default = function(object, formula, type, CIs = FALSE, PIs = FALSE,
          plotform = as.formula(paste("yvar ~ xvar |", paste(byvars, collapse="*")))
     }
     xlab = ifelse(is.null(xargs$xlab),
-                  paste("Levels of", paste(xvars, collapse=" * ")), xargs$xlab)
+                  paste0(predicate, paste(xvars, collapse=" * ")), xargs$xlab)
     rspLbl = paste("Predicted", 
                    ifelse(is.null(emmo@misc$inv.lbl), "response", emmo@misc$inv.lbl))
     ylab = ifelse(is.null(xargs$ylab),
@@ -242,17 +275,21 @@ emmip.default = function(object, formula, type, CIs = FALSE, PIs = FALSE,
         if (!is.null(col)) TP$superpose.symbol$col = TP$superpose.line$col = col
         lattice::trellis.par.set(TP)
         
+        plty = if(style=="factor") c("p","l")   else "l"
         plotspecs = list(x = plotform, data = emms, groups = ~ tvar, 
                          xlab = xlab, ylab = ylab,
-                         strip = my.strip, auto.key = my.key(tvars), type=c("p","l"))
+                         strip = my.strip, auto.key = my.key(tvars), 
+                         type = plty)
         grobj = do.call(lattice::xyplot, c(plotspecs, xargs))
         lattice::trellis.par.set(TP.orig)
     }  # --- end lattice method
     else {  # engine = "ggplot"
         pos = ggplot2::position_dodge(width = ifelse(CIs|PIs, .1, 0)) # use dodging if CIs
         if (!one.trace) {
-            grobj = ggplot2::ggplot(emms, ggplot2::aes_(x = ~xvar, y = ~yvar, color = ~tvar)) +
-                ggplot2::geom_point(position = pos) +
+            grobj = ggplot2::ggplot(emms, ggplot2::aes_(x = ~xvar, y = ~yvar, color = ~tvar))
+            if (style == "factor")
+                grobj = grobj + ggplot2::geom_point(position = pos)
+            grobj = grobj +
                 ggplot2::geom_line(ggplot2::aes_(group = ~tvar), position = pos) +
                 ggplot2::labs(x = xlab, y = ylab, color = tlab)
         }